Changed Lock to Lap + redefined
Concept of lapped brakes is the closest thing I can think of when it comes to what the Lock article is referring to. "Lock", as used in this context, is certainly not a term in western railroading afaik, especially not one used in the way the Japanese text is describing.